services/api-gateway/requirements.txt
fastapi[standard]
httpx
python-jose[cryptography]
python-dotenv